+++ acpi/Cargo.toml
@@ -12,3 +12,4 @@ edition = "2018"
 [dependencies]
 log = "0.4"
 bit_field = "0.10"
+rsdp_search = { path = "../rsdp_search" }
