+++ rsdp_search/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rsdp_search"
+version = "0.1.0"
+authors = ["Isaac Woods", "Restioson"]
+edition = "2018"
+
+[dependencies]
+log = "0.4"
