+++ rsdp_search/src/handler.rs
